Shrink dpa4 to core/interaction/so2 and unify Lebedev tables.
Move shared EMFA helpers into so3 (FocusSO2Linear, quaternion Wigner-D, envelope softmax) and models.activation SwiGLU; canonicalize Apache e3x Lebedev grids under so3 for EFA and DPA4.
