--- enerzyme/models/dpa4/data/README.md
@@ -1,6 +1,0 @@
-# Lebedev quadrature for DPA4
-#
-# The canonical copy of `lebedev_rules.npz` now lives under
-# `enerzyme/models/so3/data/` and is loaded via `enerzyme.models.so3.lebedev`.
-# This directory may retain a duplicate for package-local discovery; prefer the
-# shared so3 path for new code.
