+++ enerzyme/models/dpa4/interaction.py
@@ -1,4 +1,4 @@
-"""Interaction block for DPA4.
+"""Interaction block and equivariant FFN for DPA4.
 
 Reimplemented in PyTorch from DPA4/SeZM concepts (arXiv:2606.02419).
 """
@@ -11,10 +11,78 @@ import torch
 import torch.nn as nn
 from torch import Tensor
 
-from .edge_cache import EdgeCache
-from .ffn import EquivariantFFN
-from .norm import EquivariantRMSNorm
-from .so2 import SO2Convolution
+from ..activation import SwiGLU
+from ..so3.gated import SO3GatedActivation
+from ..so3.layer_norm import EquivariantDegreeRMSNorm
+from ..so3.lebedev import S2LebedevProjector
+from ..so3.linear import SO3FocusLinear
+from .so2 import EdgeCache, SO2Convolution
+
+
+class EquivariantFFN(nn.Module):
+    """Equivariant feed-forward network.
+
+    ``SO3FocusLinear`` → gated / Lebedev-grid SwiGLU → zero-init ``SO3FocusLinear``.
+    """
+
+    def __init__(
+        self,
+        lmax: int,
+        channels: int,
+        hidden_channels: int,
+        glu_activation: bool = True,
+        activation: str = "silu",
+        ffn_so3_grid: bool = False,
+        lebedev_quadrature: bool = True,
+    ) -> None:
+        super().__init__()
+        del lebedev_quadrature  # always Lebedev when grid path is enabled
+        self.lmax = lmax
+        self.channels = channels
+        self.hidden_channels = hidden_channels
+        self.glu_activation = glu_activation
+        self.use_grid = ffn_so3_grid
+
+        if self.use_grid:
+            self.so3_linear_1 = SO3FocusLinear(
+                lmax, channels, 2 * hidden_channels, n_focus=1
+            )
+            self.grid_proj = S2LebedevProjector(lmax)
+            self.grid_act = SwiGLU()
+        else:
+            out_ch = 2 * hidden_channels if glu_activation else hidden_channels
+            self.so3_linear_1 = SO3FocusLinear(
+                lmax, channels, out_ch, n_focus=1
+            )
+            self.act = SO3GatedActivation(
+                lmax=lmax,
+                channels=hidden_channels,
+                activation=activation,
+                layout="ndfc",
+            )
+
+        self.so3_linear_2 = SO3FocusLinear(
+            lmax, hidden_channels, channels, n_focus=1, init_std=0.0
+        )
+
+    def forward(self, x: Tensor) -> Tensor:
+        """x: (N, D, F, C) -> (N, D, F, C)."""
+        h = self.so3_linear_1(x)
+
+        if self.use_grid:
+            nc = self.hidden_channels
+            n, d, f, _ = h.shape
+            h_flat = h.reshape(n * f, d, -1)
+            g = self.grid_proj.to_grid(h_flat)
+            g = self.grid_act(g)
+            h = self.grid_proj.from_grid(g).reshape(n, d, f, nc)
+        elif self.glu_activation:
+            nc = self.hidden_channels
+            h = self.act(h[..., :nc], gate=h[..., nc:])
+        else:
+            h = self.act(h)
+
+        return self.so3_linear_2(h)
 
 
 class SeZMInteractionBlock(nn.Module):
@@ -43,7 +111,7 @@ class SeZMInteractionBlock(nn.Module):
         lebedev_quadrature: bool = True,
         glu_activation: bool = True,
         activation: str = "silu",
-        sandwich_norm: list[bool] | None = None,
+        sandwich_norm: Optional[list] = None,
         message_node_so3: bool = False,
     ) -> None:
         super().__init__()
@@ -54,20 +122,23 @@ class SeZMInteractionBlock(nn.Module):
         if sandwich_norm is None:
             sandwich_norm = [False, True, True, False]
 
-        # SO(2) pre-norm
         self.pre_so2_norm = (
-            EquivariantRMSNorm(lmax, channels)
-            if sandwich_norm[0] else nn.Identity()
+            EquivariantDegreeRMSNorm(lmax, channels)
+            if sandwich_norm[0]
+            else nn.Identity()
         )
-        # SO(2) post-norm
         self.post_so2_norm = (
-            EquivariantRMSNorm(lmax, channels)
-            if sandwich_norm[1] else nn.Identity()
+            EquivariantDegreeRMSNorm(lmax, channels)
+            if sandwich_norm[1]
+            else nn.Identity()
         )
 
         self.so2_conv = SO2Convolution(
-            lmax=lmax, mmax=mmax, channels=channels,
-            n_focus=n_focus, focus_dim=focus_dim,
+            lmax=lmax,
+            mmax=mmax,
+            channels=channels,
+            n_focus=n_focus,
+            focus_dim=focus_dim,
             mixing_layers=mixing_layers,
             n_atten_head=n_atten_head,
             radial_so2_mode=radial_so2_mode,
@@ -78,27 +149,31 @@ class SeZMInteractionBlock(nn.Module):
             message_node_so3=message_node_so3,
         )
 
-        # FFN subblocks
         self.pre_ffn_norms = nn.ModuleList()
         self.post_ffn_norms = nn.ModuleList()
         self.ffns = nn.ModuleList()
         for _ in range(ffn_blocks):
             self.pre_ffn_norms.append(
-                EquivariantRMSNorm(lmax, channels)
-                if sandwich_norm[2] else nn.Identity()
+                EquivariantDegreeRMSNorm(lmax, channels)
+                if sandwich_norm[2]
+                else nn.Identity()
             )
             self.post_ffn_norms.append(
-                EquivariantRMSNorm(lmax, channels)
-                if sandwich_norm[3] else nn.Identity()
+                EquivariantDegreeRMSNorm(lmax, channels)
+                if sandwich_norm[3]
+                else nn.Identity()
             )
-            self.ffns.append(EquivariantFFN(
-                lmax=lmax, channels=channels,
-                hidden_channels=ffn_neurons,
-                glu_activation=glu_activation,
-                activation=activation,
-                ffn_so3_grid=ffn_so3_grid,
-                lebedev_quadrature=lebedev_quadrature,
-            ))
+            self.ffns.append(
+                EquivariantFFN(
+                    lmax=lmax,
+                    channels=channels,
+                    hidden_channels=ffn_neurons,
+                    glu_activation=glu_activation,
+                    activation=activation,
+                    ffn_so3_grid=ffn_so3_grid,
+                    lebedev_quadrature=lebedev_quadrature,
+                )
+            )
 
     def forward(self, x: Tensor, edge_cache: EdgeCache, radial_feat: Tensor) -> Tensor:
         """
@@ -110,20 +185,11 @@ class SeZMInteractionBlock(nn.Module):
         Returns:
             (N, D, 1, C) updated features
         """
-        N, D, _, C = x.shape
-
-        # SO(2) branch
-        x_so2 = x
-        x_pre = self.pre_so2_norm(x_so2)
-        so2_out = self.so2_conv(
-            x_pre.squeeze(2),  # (N, D, C)
-            edge_cache,
-            radial_feat,
-        )
-        so2_out = self.post_so2_norm(so2_out.unsqueeze(2))  # (N, D, 1, C)
+        x_pre = self.pre_so2_norm(x)
+        so2_out = self.so2_conv(x_pre.squeeze(2), edge_cache, radial_feat)
+        so2_out = self.post_so2_norm(so2_out.unsqueeze(2))
         x = x + so2_out
 
-        # FFN branch(es)
         for i in range(self.ffn_blocks_count):
             x_ffn = self.pre_ffn_norms[i](x)
             y = self.ffns[i](x_ffn)
