--- enerzyme/models/dpa4/radial.py
@@ -1,10 +1,0 @@
-"""Radial basis for DPA4 — re-exports shared C³ Bessel utilities."""
-
-from __future__ import annotations
-
-from ..so3 import C3CutoffEnvelope
-from ..so3.radial import BesselC3RadialBasis, RadialMLP
-
-RadialBasis = BesselC3RadialBasis
-
-__all__ = ["BesselC3RadialBasis", "C3CutoffEnvelope", "RadialBasis", "RadialMLP"]
