+++ enerzyme/models/dpa4/so2.py
@@ -1,12 +1,13 @@
-"""SO(2)-equivariant convolution for DPA4.
+"""SO(2)-equivariant convolution for DPA4 EMFA.
 
-Reimplemented in PyTorch from DPA4/SeZM concepts (arXiv:2606.02419).
-Supports degree_channel radial mixing, multi-head attention, gated activation.
+EdgeCache, DynamicRadialDegreeMixer, and SO2Convolution stay local.
+Focus-major SO(2) linear lives in ``enerzyme.models.so3.so2_focus``.
 """
 
 from __future__ import annotations
 
 import math
+from dataclasses import dataclass
 from typing import List, Optional, Tuple
 
 import numpy as np
@@ -14,10 +15,8 @@ import torch
 import torch.nn as nn
 from torch import Tensor
 
-from .activation import GatedActivation
-from .attention import segment_envelope_gated_softmax
-from .edge_cache import EdgeCache
-from .indexing import (
+from ..so3.gated import SO3GatedActivation
+from ..so3.indexing import (
     build_m_major_index,
     build_m_major_l_index,
     build_rotate_inv_rescale,
@@ -25,100 +24,77 @@ from .indexing import (
     project_D_to_m,
     project_Dt_from_m,
 )
+from ..so3.so2_focus import FocusSO2Linear
+from ..so3.softmax import segment_envelope_gated_softmax
+from ..so3.wigner_quaternion import (
+    WignerDCalculator,
+    build_edge_quaternion,
+    quaternion_multiply,
+    quaternion_z_rotation,
+    safe_norm,
+)
 
 
-class SO2Linear(nn.Module):
-    """SO(2)-equivariant linear in m-major reduced layout.
+@dataclass
+class EdgeCache:
+    """Per-edge geometry cache created once per forward pass."""
 
-    Weight is block-diagonal over |m| groups:
-    - m=0: unconstrained cross-l mixing
-    - |m|>0: SO(2)-constrained 2x2 coupling of (-m, +m) pairs
-    """
+    src: Tensor
+    dst: Tensor
+    edge_vec: Tensor
+    edge_rbf: Tensor
+    edge_env: Tensor
+    deg: Tensor
+    inv_sqrt_deg: Tensor
+    D_full: Optional[Tensor] = None
+    Dt_full: Optional[Tensor] = None
+    edge_quat: Optional[Tensor] = None
 
-    def __init__(
-        self,
-        lmax: int,
-        mmax: int,
-        in_channels: int,
-        out_channels: int,
-        n_focus: int = 1,
-        bias: bool = False,
-    ) -> None:
-        super().__init__()
-        self.lmax = lmax
-        self.mmax = mmax
-        self.in_channels = in_channels
-        self.out_channels = out_channels
-        self.n_focus = n_focus
-        self.use_bias = bias
 
-        num_l_m0 = lmax + 1
-        self.weight_m0 = nn.Parameter(
-            torch.empty(n_focus, num_l_m0 * in_channels, num_l_m0 * out_channels)
-        )
-        nn.init.xavier_uniform_(self.weight_m0)
+def build_edge_cache(
+    *,
+    idx_i: Tensor,
+    idx_j: Tensor,
+    vij: Tensor,
+    n_nodes: int,
+    radial_basis: torch.nn.Module,
+    envelope: torch.nn.Module,
+    wigner_calc: WignerDCalculator,
+    random_gamma: bool = False,
+    deg_norm_floor: float = 1e-7,
+) -> EdgeCache:
+    """Build the edge cache from a sparse edge list."""
+    device = vij.device
+    dtype = vij.dtype
+    src = idx_i.long()
+    dst = idx_j.long()
 
-        if bias:
-            self.bias0 = nn.Parameter(torch.zeros(n_focus, out_channels))
-        else:
-            self.bias0 = None
+    edge_len = safe_norm(vij)
+    edge_env = envelope(edge_len)
+    edge_rbf = radial_basis(edge_len)
 
-        self.weight_m = nn.ParameterList()
-        for m in range(1, mmax + 1):
-            num_l = lmax - m + 1
-            w = nn.Parameter(
-                torch.empty(n_focus, num_l * in_channels, 2 * num_l * out_channels)
-            )
-            nn.init.xavier_uniform_(w)
-            w.data *= 1.0 / math.sqrt(2.0)
-            self.weight_m.append(w)
+    edge_quat = build_edge_quaternion(vij, edge_len)
+    if random_gamma and wigner_calc.training:
+        gamma = torch.rand(src.shape[0], device=device, dtype=dtype) * (2 * math.pi)
+        edge_quat = quaternion_multiply(quaternion_z_rotation(gamma), edge_quat)
+    D_full, Dt_full = wigner_calc(edge_quat)
 
-        self.reduced_dim = (lmax + 1) + sum(
-            2 * (lmax - m + 1) for m in range(1, mmax + 1)
-        )
+    deg = torch.zeros(n_nodes, device=device, dtype=dtype)
+    deg.scatter_add_(0, dst, (edge_env.squeeze(-1) ** 2))
+    inv_sqrt_deg = (1.0 / torch.sqrt(deg + deg_norm_floor)).reshape(n_nodes, 1, 1)
 
-    def forward(self, x: Tensor) -> Tensor:
-        """x: (F, E, D_m_trunc, Cin) -> (F, E, D_m_trunc, Cout)."""
-        F, E = x.shape[0], x.shape[1]
-        num_l_m0 = self.lmax + 1
-
-        x_m0 = x[:, :, :num_l_m0, :]
-        x_m0_flat = x_m0.reshape(F, E, num_l_m0 * self.in_channels)
-        out_m0 = torch.bmm(x_m0_flat, self.weight_m0)
-        out_m0 = out_m0.reshape(F, E, num_l_m0, self.out_channels)
-
-        if self.use_bias and self.bias0 is not None:
-            out_m0 = torch.cat(
-                [
-                    out_m0[:, :, :1, :] + self.bias0[:, None, None, :],
-                    out_m0[:, :, 1:, :],
-                ],
-                dim=2,
-            )
-
-        blocks = [out_m0]
-        offset = num_l_m0
-        for m_idx, m in enumerate(range(1, self.mmax + 1)):
-            num_l = self.lmax - m + 1
-            x_neg = x[:, :, offset : offset + num_l, :]
-            x_pos = x[:, :, offset + num_l : offset + 2 * num_l, :]
-            offset += 2 * num_l
-
-            x_neg_flat = x_neg.reshape(F, E, num_l * self.in_channels)
-            x_pos_flat = x_pos.reshape(F, E, num_l * self.in_channels)
-
-            w = self.weight_m[m_idx]
-            w_u = w[:, :, : num_l * self.out_channels]
-            w_v = w[:, :, num_l * self.out_channels :]
-
-            # Complex multiply (u+iv): out_neg = a u - b v, out_pos = a v + b u
-            out_neg = torch.bmm(x_neg_flat, w_u) - torch.bmm(x_pos_flat, w_v)
-            out_pos = torch.bmm(x_neg_flat, w_v) + torch.bmm(x_pos_flat, w_u)
-
-            blocks.append(out_neg.reshape(F, E, num_l, self.out_channels))
-            blocks.append(out_pos.reshape(F, E, num_l, self.out_channels))
-
-        return torch.cat(blocks, dim=2)
+    return EdgeCache(
+        src=src,
+        dst=dst,
+        edge_vec=vij,
+        edge_rbf=edge_rbf,
+        edge_env=edge_env,
+        deg=deg,
+        inv_sqrt_deg=inv_sqrt_deg,
+        D_full=D_full,
+        Dt_full=Dt_full,
+        edge_quat=edge_quat,
+    )
 
 
 class DynamicRadialDegreeMixer(nn.Module):
@@ -371,7 +347,7 @@ class SO2Convolution(nn.Module):
         self.so2_acts = nn.ModuleList()
         for _ in range(mixing_layers):
             self.so2_layers.append(
-                SO2Linear(
+                FocusSO2Linear(
                     lmax=lmax,
                     mmax=mmax,
                     in_channels=self.Cf,
@@ -380,7 +356,7 @@ class SO2Convolution(nn.Module):
                 )
             )
             self.so2_acts.append(
-                GatedActivation(
+                SO3GatedActivation(
                     lmax=lmax,
                     channels=self.Cf,
                     n_focus=n_focus,
