+++ enerzyme/models/efa/NOTICE
@@ -1,8 +1,9 @@
 Lebedev quadrature tables
 -------------------------
-``lebedev_grids.npz`` is copied from Google e3x
+``enerzyme/models/so3/data/lebedev_grids.npz`` is copied from Google e3x
 (https://github.com/google-research/e3x, file ``e3x/so3/_lebedev_grids.npz``)
-licensed under the Apache License, Version 2.0.
+licensed under the Apache License, Version 2.0. EFA loads it via
+``enerzyme.models.so3.lebedev`` (re-exported from ``enerzyme.models.efa.lebedev``).
 
 ERoPE / linear EFA aggregation
 ------------------------------
