+++ enerzyme/models/efa/__init__.py
@@ -9,8 +9,8 @@ Architecture-agnostic nonlocal plug-in for Enerzyme Cores:
 * **Other Cores** — call :class:`EFABlock` or :func:`apply_efa_if_configured`
   on invariant ``[N, F]`` features with absolute ``Ra`` and ``batch_seg``.
 
-Lebedev grids are vendored from Google e3x (Apache-2.0); see
-``lebedev_grids.npz`` and ``NOTICE``.
+Lebedev grids are the shared e3x Apache-2.0 tables under
+``enerzyme.models.so3.data`` (re-exported here; see ``NOTICE``).
 """
 
 from .attention import EuclideanFastAttention
