+++ enerzyme/models/so3/data/lebedev_grids.npz
