+++ setup.py
@@ -20,7 +20,7 @@ setup(
         "models/layers/dispersion/grimme_d4_tables/*.pth",
         "models/so3/Jd.pt",
         "models/so3/cgmatrix.npz",
-        "models/efa/lebedev_grids.npz",
+        "models/so3/data/lebedev_grids.npz",
         "models/efa/NOTICE",
     ]},
     auth='Benzoin96485',
